first attempt at reorganization and refactoring to make a library and program
