+++ Makefile
@@ -1,4 +1,5 @@
 unsf:
+	cc -g -O2 -o libunsf.o libunsf.c
 	cc -g -O2 -o unsf unsf.c -lm
 install: unsf
 	install unsf $(DESTDIR)/usr/bin/
